idea-agent/requirements.txt
fastapi
uvicorn
requests
langchain
langchain-community
langgraph
langchain-openai
langchain-anthropic
redis
mysql-connector-python
python-dotenv
